+++ src/ServiceCluster.ts
@@ -52,8 +52,6 @@ export class ServiceCluster implements IClassServiceCluster {
   public async update(arg: IServiceClusterUpdate) {
     if (!arg)
       throw new Error(`serviceCluster.update: all arguments are missing`);
-    if (arg && !arg.id)
-      throw new Error(`serviceCluster.update: "id" parameter is required`);
 
     if (arg.name) this.details.name = arg.name;
     if (arg.persistenceMode)
@@ -88,7 +86,7 @@ export class ServiceCluster implements IClassServiceCluster {
         arg.failoverTimeout;
 
     return await this.repoClient
-      .Post(`ServiceCluster/${arg.id}`, { ...this.details })
+      .Post(`ServiceCluster/${this.details.id}`, { ...this.details })
       .then((res) => res.status);
   }
 }